refactor: move old payload classes
