+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/injection/PackageVersionInfo.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.payload
+package io.embrace.android.embracesdk.internal.injection
 
 import android.content.pm.PackageInfo
