+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/session/message/FinalEnvelopeParams.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.internal.session.message
 
 import io.embrace.android.embracesdk.internal.logging.EmbLogger
-import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import io.embrace.android.embracesdk.internal.session.SessionZygote
 import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSnapshotType
 
 /**
