+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/session/orchestrator/SessionSpanAttrPopulatorImplTest.kt
@@ -7,7 +7,7 @@ import io.embrace.android.embracesdk.fakes.FakeMetadataService
 import io.embrace.android.embracesdk.fakes.FakeSessionPropertiesService
 import io.embrace.android.embracesdk.internal.payload.ApplicationState
 import io.embrace.android.embracesdk.internal.payload.LifeEventType
-import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import io.embrace.android.embracesdk.internal.session.SessionZygote
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
