+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakePayloadCachingService.kt
@@ -2,7 +2,7 @@ package io.embrace.android.embracesdk.fakes
 
 import io.embrace.android.embracesdk.internal.delivery.caching.PayloadCachingService
 import io.embrace.android.embracesdk.internal.delivery.caching.SessionPayloadSupplier
-import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import io.embrace.android.embracesdk.internal.session.SessionZygote
 import io.embrace.android.embracesdk.internal.session.lifecycle.ProcessState
 
 class FakePayloadCachingService : PayloadCachingService {
