+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeSessionSpanAttrPopulator.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.fakes
 
 import io.embrace.android.embracesdk.internal.payload.LifeEventType
-import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import io.embrace.android.embracesdk.internal.session.SessionZygote
 import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSpanAttrPopulator
 
 class FakeSessionSpanAttrPopulator : SessionSpanAttrPopulator {
